requirements.txt
imagehash==4.3.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
monai==1.5.0
networkx==3.5
numpy==2.2.6
opencv_python==4.12.0.88
pandas==2.3.0
Pillow==11.2.1
pytorch_lightning==2.5.2
PyYAML==6.0.2
scikit_learn==1.7.0
scipy==1.15.3
segmentation_models_pytorch==0.5.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
